+++ src/components/App/Apps/GuideShow/Panels/LeftPanel/Settings/PrivacySecurity/Status.tsx
@@ -1,0 +1,17 @@
+import styled from "styled-components"
+import Header from "../../Header"
+import React from 'react'
+
+const StyledContainer = styled.div`
+  height: 100%;
+  width: 100%;
+`
+
+const Status = () => {
+    return (
+        <StyledContainer>
+            <Header/>
+        </StyledContainer>
+    )
+}
+export default Status